Refactored DeleteComment code to only allow deletion of comments written by the current logged in user

+++ northcoders---frontend---news-api/src/Components/DeleteComment.jsx
@@ -1,13 +1,15 @@
 import React from "react";
 import * as api from "../utils/api";
 
-const DeleteComment = ({ comment_id, getComments }) => {
+const DeleteComment = ({ comment_id, getComments, author, username }) => {
 
   const handleSubmit = event => {
-    event.preventDefault();
-    api.deleteComment(comment_id).then(() => {
-      getComments();
-    });
+    event.preventDefault(); 
+    if(author === username) {
+      api.deleteComment(comment_id).then(() => {
+        getComments();
+      });  
+    }       
   };
   return (
     <div id='DeleteButton'>
